Use separate Gson for MS stats serialization/deserialization
